% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countryGroup.map}
\alias{countryGroup.map}
\title{Internal function to Map Countries in IMF Country Categorizations to Columns
in Country Data in IMF WEO forecast data}
\usage{
countryGroup.map(countryGroups, WEOcountries, groupNames, altNames = NULL)
}
\arguments{
\item{countryGroups}{data.table containing}

\item{WEOcountries}{vector of countries from WEO data}

\item{groupNames}{which of the columns in CountryGroups should be mapped to
new variable}

\item{altNames}{optional; alternative Names for the groups (defaults to using
groupNames)}
}
\value{
a vector with country categorization
}
\description{
Internal function to Map Countries in IMF Country Categorizations to Columns
in Country Data in IMF WEO forecast data
}
