% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{download.data}
\alias{download.data}
\title{Downloads WEO forecast data from the internet}
\usage{
download.data()
}
\value{
nothing, downs
}
\description{
Downloads WEO forecast data from the internet
}
\examples{
download.data()
}
