% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{download.process.weo}
\alias{download.process.weo}
\title{Function to download, tidy and save WEO data}
\usage{
download.process.weo(
  sheets = c("ngdp_rpch", "pcpi_pch"),
  explicitMissings = TRUE,
  target_filename = "WEOforecasts_tidy.csv",
  includeCountryGroups = TRUE,
  fileCountryCat = "FMEconGroup.xlsx"
)
}
\arguments{
\item{sheets}{which sheets to read in. One (or more) of "ngdp_rpch",
"pcpi_pch", "bca_gdp_bp6". Default: "ngdp_rpch", "pcpi_pch" (Inflation and
GDP growth)}

\item{explicitMissings}{should missing prediction be explicitly included
(coded as NA values), default: TRUE}

\item{target_filename}{name of the target csv file where tidied data will
be stored}
}
\value{
nothing, saves tidied data in directory
}
\description{
Function to download, tidy and save WEO data
}
