% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imfpp-package.R
\docType{package}
\name{imfpp-package}
\alias{imfpp}
\alias{imfpp-package}
\title{imfpp: Download and Tidy IMF Forecast Data}
\description{
Convenience functions to download and tidy IMF forecast data for GDP growth, inflation and unemployment.
}
\author{
\strong{Maintainer}: Becker Rike \email{f-becker@posteo.de}

}
\keyword{internal}
