% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{incl.country.cat}
\alias{incl.country.cat}
\title{This is a function to clean some minor details in the WEO forecasts data}
\usage{
incl.country.cat(tidiedWEO, fileCountryCat = "FMEconGroup.xlsx")
}
\arguments{
\item{tidiedWEO}{tidied WEO data}

\item{fileCountryCat}{"FMEconGroup.xlsx"}
}
\value{
data.table with country categorization
}
\description{
This is a function to clean some minor details in the WEO forecasts data
}
