% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read.sheet}
\alias{read.sheet}
\title{Title}
\usage{
read.sheet(sheetname, explicitMissings = TRUE, filename = "WEOforecasts.xlsx")
}
\arguments{
\item{sheetname}{name of sheet}

\item{explicitMissings}{should missings be explicit, default: TRUE}

\item{filename}{File name of Excel file}
}
\value{
data.table object
}
\description{
Title
}
\examples{
read.sheet("ngdp_rpch")

}
