% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read.weodata}
\alias{read.weodata}
\title{Simple wrapper to bind rows for multiple targets in IMF WEO Data}
\usage{
read.weodata(
  sheets = c("ngdp_rpch", "pcpi_pch"),
  explicitMissings = TRUE,
  filename = "WEOforecasts.xlsx"
)
}
\arguments{
\item{sheets}{which sheets to read in. One (or more) of "ngdp_rpch",
"pcpi_pch", "bca_gdp_bp6"}

\item{explicitMissings}{should missing prediction be explicitly included
(coded as NA values), default: TRUE}

\item{filename}{name of the Excel file containing the WEO forecasts}
}
\value{
a data.table object of tidied and forecast data of the specified targets
}
\description{
Simple wrapper to bind rows for multiple targets in IMF WEO Data
}
\examples{
read.weodata(sheets = c("pcpi_pch"))
}
