% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{tidy.data}
\alias{tidy.data}
\title{Function to tidy the data that come from the IMF WEO historical forecast
Excel sheets}
\usage{
tidy.data(weoSheet)
}
\arguments{
\item{weoSheet}{}
}
\value{
tidied data.table
}
\description{
Function to tidy the data that come from the IMF WEO historical forecast
Excel sheets
}
\examples{
tidy.data(weoSheet)
}
